Adds buttons to choose tool, removes Text UI, removes Panel Button Handler and moves code to UI handler
