--- Assets/Scripts/CurrentToolTextController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 988bcdc98b965e747b118e07f846697a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
