--- Assets/Scripts/UI Scripts/PanelButtonHandler.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b7431c7df7835ad4299c04efe594d3a6
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
