+++ Assets/Scripts/UI Scripts/UIHandler.cs
@@ -1,34 +1,71 @@
-﻿using System.Collections;
+﻿using System;
+using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.SceneManagement;
 
 public class UIHandler : MonoBehaviour
 {
-    public void RestartOnClick()
+
+    //observer pattern that tells current tool UI what tool is chosen;
+    public static event Action<int> CurrentToolTriggered;
+
+    //opens and closes UI panel
+    public void OpenPanel() 
+    {
+        Animator animator = gameObject.GetComponent<Animator>();
+        bool isOpen = animator.GetBool("is Open");
+
+        animator.SetBool("is Open", !isOpen);
+    }
+
+    //restarts level
+    public void RestartOnClick() 
     {
         SceneManager.LoadScene(0);
     }
 
-    public void PlayOnClick() //starts game and sets time to normal
+    //starts game and sets time to normal
+    public void PlayOnClick() 
     {
         Time.timeScale = 1;
         Time.fixedDeltaTime = 0.02f;
     }
 
-    public void PauseOnClick() //pauses game movement
+    //pauses game movement
+    public void PauseOnClick() 
     {
         Time.timeScale = 0;
     }
 
-    public void SlowOnClick() //sets game to slow motion
+    //sets game to slow motion
+    public void SlowOnClick() 
     {
         Time.timeScale = 0.5f;
         Time.fixedDeltaTime = 0.02f * Time.timeScale;
     }
 
-    public void FastOnClick() //sets game to 2x the speed
+    //sets game to 2x the speed
+    public void FastOnClick() 
     {
         Time.timeScale = 2;
+    }
+
+    //sets current tool to spring
+    public void SpringOnClick()
+    {
+        if (CurrentToolTriggered != null)
+        {
+            CurrentToolTriggered.Invoke(0);
+        }
+    }
+
+    //sets current tool to wall
+    public void WallOnClick()
+    {
+        if (CurrentToolTriggered != null)
+        {
+            CurrentToolTriggered.Invoke(1);
+        }
     }
 }
